Moved storageaccount property check function to azure utils
